requirements.txt
# =============================================================================
# FRGT Requirements
# =============================================================================
# Dependencies for Flow Reconstructing Graph Transformer (FRGT)
# This file lists all required packages for training and inference

# -----------------------------------------------------------------------------
# Core Deep Learning Framework
# -----------------------------------------------------------------------------
torch>=2.0.0                    # PyTorch deep learning framework
torchvision>=0.15.0             # Computer vision utilities (for data loading)
torchaudio>=2.0.0               # Audio processing (PyTorch ecosystem)

# -----------------------------------------------------------------------------  
# Graph Neural Networks
# -----------------------------------------------------------------------------
torch-geometric>=2.4.0         # Graph neural network library
torch_cluster>=1.6.0           # Graph clustering algorithms
torch_scatter>=2.1.0           # Efficient scatter operations for graphs
torch_sparse>=0.6.15           # Sparse tensor operations
torch_spline_conv>=1.2.0       # Spline convolution operations
pyg_lib>=0.3.0                 # PyTorch Geometric extensions

# -----------------------------------------------------------------------------
# Scientific Computing & Data Processing
# -----------------------------------------------------------------------------
numpy>=1.21.0                  # Fundamental array computing
scipy>=1.7.0                   # Scientific computing algorithms
pandas>=1.3.0                  # Data manipulation and analysis
networkx>=2.6                  # Network analysis and graph algorithms

# -----------------------------------------------------------------------------
# Configuration & Utilities
# -----------------------------------------------------------------------------
python-box>=6.0.0              # Box notation for configuration files
pyyaml>=6.0                    # YAML configuration file parsing
tqdm>=4.60.0                   # Progress bars for training loops

# -----------------------------------------------------------------------------
# Visualization & Analysis
# -----------------------------------------------------------------------------
matplotlib>=3.5.0              # Plotting and visualization
seaborn>=0.11.0                # Statistical data visualization

# -----------------------------------------------------------------------------
# CFD & Scientific Data Processing
# -----------------------------------------------------------------------------
fluidfoam>=0.2.5               # OpenFOAM data processing utilities
netCDF4>=1.5.0                 # NetCDF file format support
PyFoam>=2022.0                 # OpenFOAM Python utilities

# -----------------------------------------------------------------------------
# Jupyter & Development (Optional)
# -----------------------------------------------------------------------------
jupyter>=1.0.0                 # Jupyter notebook environment
ipython>=7.0.0                 # Interactive Python shell
ipykernel>=6.0.0               # Jupyter kernel for Python

# -----------------------------------------------------------------------------
# Additional Dependencies
# -----------------------------------------------------------------------------
requests>=2.25.0               # HTTP library for data downloading
pillow>=8.0.0                  # Image processing capabilities
